requirements.txt
flask
opennsfw2
Pillow
opencv-python